requirements.txt
kafka-python
pandas
sqlalchemy
psycopg2